requirements.txt
# Python >= 3.9
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
python-dotenv
faiss-cpu
openai
pydantic
typing-extensions
python-multipart
pytest